requirements.txt
ewstools==2.1.2
ipython==8.12.3
matplotlib==3.7.1
networkx==3.4.2
numpy==2.4.2
pandas==1.5.3
python_igraph==1.0.0
PyYAML==6.0.3
scikit_learn==1.3.0
scipy==1.17.1
seaborn==0.13.2
torch==2.0.1
torch_geometric==2.7.0
torch_sparse==0.6.17+pt20cu118
torch_timeseries==0.1.10
torchsde==0.2.6
tqdm==4.67.1
